refactor(domains): move cart-store to domains/orders (#232)
